refactor: Renamce deploy.ts to config.ts
