+++ packages/secret-contracts-scripts/src/config/config.ts
