+++ packages/secret-contracts-scripts/src/instantiateGateway.ts
@@ -1,7 +1,7 @@
 import { BroadcastMode, SecretNetworkClient, Wallet } from "secretjs";
 import path from 'path';
 import { fileURLToPath } from 'url';
-import config from './config/deploy.js';
+import config from './config/config.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
